% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfou.R
\name{acf_fou}
\alias{acf_fou}
\title{Autocovariance function of the fOU}
\usage{
acf_fou(a, nu, H, lag_max = 50, delta = 1/252)
}
\arguments{
\item{a}{Speed of mean reversion coefficient.}

\item{nu}{Diffusion coefficient.}

\item{H}{Hurst coefficient.}

\item{lag_max}{Integer. The autocovariance is calculated for all integers between
\eqn{0} and \code{lag_max}.}

\item{delta}{Time step in the discrete representation of the process (e.g., 1/252 for daily observations if time is in years).}
}
\value{
Numeric vector.
}
\description{
These functions calculate the autocovariance function of the fractional
Ornstein-Uhlenbeck process for a number of lags, based on on model parameters.
\itemize{
\item \code{acf_fou} calculates the exact autocovariances;
\item \code{acf_asy_fou}calculates the asymptotic autocovariances arising in the regime of
small speed of mean reversion.
}
}
