% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bold}
\alias{bold}
\title{BOLD signal}
\format{
A data frame with 3600 rows and 16 columns:
\describe{
\item{HIP}{Hippocampus.}
\item{AMY}{Amygdala.}
\item{pTHA}{Posterio Thalamus.}
\item{aTHA}{Anterior Thalamus.}
\item{NAc}{Nucleus Accumbens.}
\item{GP}{Globus Pallidus.}
\item{PUT}{Putamen.}
\item{CAU}{Caudate.}
\item{HIP}{Hippocampus.}
}
}
\source{
Data obtained from the Padova Neuroscience Center.
}
\usage{
bold
}
\description{
Per-second observations of a healthy individual's 7T resting-state fMRI, drawn from the Human Connectome Project,
based on Tian’s subcortical parcellations. RH and LH stand for right and left hemisphere,
respectively.
}
\keyword{datasets}
