% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfou.R
\name{ccf_mfou}
\alias{ccf_mfou}
\title{Cross-covariance function of the mfOU}
\usage{
ccf_mfou(ai, aj, nui, nuj, Hi, Hj, rhoij, etaij, lag_max = 50, delta = 1/252)
}
\arguments{
\item{ai}{Speed of mean reversion of component i.}

\item{aj}{Speed of mean reversion of component j.}

\item{nui}{Diffusion coefficient of component i.}

\item{nuj}{Diffusion coefficient of component j.}

\item{Hi}{Hurst coefficient of component i.}

\item{Hj}{Hurst coefficient of component j.}

\item{rhoij}{Correlation coefficient of the underlying mfBm.}

\item{etaij}{Asymmetry coefficient of the underlying mfBm.}

\item{lag_max}{Integer. The cross-covariance is calculated for all integers between
\eqn{0} and \code{lag_max}.}

\item{delta}{Time step in the discrete representation of the process (e.g., 1/252 for daily observations if time is in years).}
}
\value{
Numeric value.
}
\description{
These functions calculate the cross-covariance function of the multivariate
fractional Ornstein-Uhlenbeck process for a number of lags, based on on model parameters.
\itemize{
\item \code{ccf_mfou} calculates the exact cross-covariances;
\item \code{ccf_asy_mfou}calculates the asymptotic cross-covariances arising in the regime of
small speed of mean reversion coefficients.
}
}
