% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfou.R
\name{cov_iik}
\alias{cov_iik}
\alias{cov_ijk}
\title{Sample cross-covariance estimator}
\usage{
cov_iik(yi, k, mui = NULL)

cov_ijk(yi, yj, k, mui = NULL, muj = NULL)
}
\arguments{
\item{yi}{Numeric vector representing a time series.}

\item{k}{Lag (integer) at which to calculate the cross-covariance.}

\item{mui}{Mean of yi. If NULL, the sample average is used.}

\item{yj}{Numeric vector of the same length of yi representing a time series.}

\item{muj}{Mean of yj. If NULL, the sample average is used.}
}
\value{
Numeric value.
}
\description{
This function calculates the sample estimate of the cross-covariance from
observations:
\itemize{
\item \code{cov_iik} calculates the sample autocovariance at lag k for a time series \eqn{y^i},
\deqn{\frac{1}{n-k}\sum_{t=1}^{n-k} y^i_{t+k} y^i_{t}-\mu_i^2;}
\item \code{cov_ijk} calculates the sample cross-covariance between two time series of
the same length, \eqn{y^i} at time t + k and \eqn{y^j} at time t,
\deqn{\frac{1}{n-k}\sum_{t=1}^{n-k} y^i_{t+k} y^j_{t}-\mu_i\mu_j,}
\eqn{k\in\mathbb{Z}}.
}
}
\examples{
x = sin(1 : 100) + rnorm(1 : 100)
y = cos(1 : 100) + rnorm(1 : 100)
cov_iik(y, 0)
cov_ijk(x, y, 0)
cov_ijk(x, y, 1)
}
