% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfbm.R
\name{cov_existence_mfbm}
\alias{cov_existence_mfbm}
\alias{cov_existence_2fbm}
\title{Existence condition of the covariance matrix of the mfBm}
\usage{
cov_existence_mfbm(H, rho, eta, out = FALSE)

cov_existence_2fbm(rhoij, etaij, Hi, Hj, out = FALSE)
}
\arguments{
\item{H}{Vector of Hurst coefficients.}

\item{rho}{Symmetric matrix of instantaneous correlation coefficients.}

\item{eta}{Antisymmetric matrix of asymmetry coefficients.}

\item{out}{Logical. If TRUE, the eigenvalues of the auxiliary matrix are
returned.}

\item{rhoij}{Instantaneous correlation coefficient in the 2-dimensional case.}

\item{etaij}{Asymmetry coefficient in the 2-dimensional case.}

\item{Hi}{Hurst coefficient of first component in the 2-dimensional case.}

\item{Hj}{Hurst coefficient of second component in the 2-dimensional case.}
}
\value{
Both functions print a message regarding the admissibility of the parameters.
\code{cov_existence_2fbm} returns the value of the constrains, whereas \code{cov_existence_mfbm} returns the
eigenvalues of the auxiliary matrix.
}
\description{
Functions to verify the existence of the covariance matrix of a multivariate
fractional Brownian motion given parameters' values.
\itemize{
\item \code{cov_existence_2fbm} verifies the simplified condition arising in the
two dimensional case, referred to in the paper as coherency constraint;
\item \code{cov_existence_mfbm} verifies the general existence condition in terms
of the positive semi-definiteness of an auxiliary matrix.
}
}
\examples{
H = c(0.0934, 0.1345, 0.1490, 0.1664)
rho = matrix(c(1.0000, 0.1471, 0.3143, 0.0739,
               0.1471, 1.0000, 0.0599, 0.1218,
               0.3143, 0.0599, 1.0000, 0.0299,
               0.0739, 0.1218, 0.0299, 1.0000),
             ncol = 4)
eta = matrix(c(0, 0.0736, -0.1752, 0.0318,
               -0.0736, 0, - 0.1550, - 0.0073,
               0.1752, 0.1550, 0, 0.0522,
               -0.0318, 0.0073, -0.0522, 0),
             ncol = 4)

# the parameters are admissible for a 4-dimensional mfBm
cov_existence_mfbm(H, rho, eta)

# the parameters are not always admissible for 2-dimensional mfBms
for(i in 1 : 4){
 for(j in 1 : i) {
   cov_existence_2fbm(rho[i, j], eta[i, j], H[i], H[j])
 }
}
}
\references{
Amblard, Pierre-Olivier, et al. "Basic properties of the multivariate fractional Brownian motion." arXiv preprint arXiv:1007.0828 (2010).
}
