% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfou.R
\name{cov_iik_mfou}
\alias{cov_iik_mfou}
\alias{cov_ij0_mfou}
\alias{cov_ijk_mfou}
\alias{cov_jik_mfou}
\title{Cross-covariance function of the mfOU process}
\usage{
cov_iik_mfou(ai, nui, Hi, k)

cov_ij0_mfou(ai, aj, nui, nuj, Hi, Hj, rhoij, etaij)

cov_ijk_mfou(ai, aj, nui, nuj, Hi, Hj, rhoij, etaij, k)

cov_jik_mfou(ai, aj, nui, nuj, Hi, Hj, rhoij, etaij, k)
}
\arguments{
\item{ai}{Speed of mean reversion of component i.}

\item{nui}{Diffusion coefficient of component i.}

\item{Hi}{Hurst coefficient of component i.}

\item{k}{Lag, \eqn{k >= 0}.}

\item{aj}{Speed of mean reversion of component j.}

\item{nuj}{Diffusion coefficient of component j.}

\item{Hj}{Hurst coefficient of component j.}

\item{rhoij}{Correlation coefficient of the underlying mfBm.}

\item{etaij}{Asymmetry coefficient of the underlying mfBm.}
}
\value{
Numeric value.
}
\description{
This function calculates the covariance function of the multivariate fractional
Ornstein-Uhlenbeck (mfOU) process, based on model parameters:
\itemize{
\item \code{cov_iik_mfou} calculates the autocovariance at lag k of component i,
\deqn{\text{Cov}\left(Y_{t+k}^i,Y_t^i\right);}
\item \code{cov_ij0_mfou} calculates the covariance between component i and j,
\deqn{\text{Cov}\left(Y_t^i,Y_t^j\right);}
\item \code{cov_ijk_mfou} calculates the covariance between component i at time \eqn{t+k} and
component j at time \eqn{t}, \deqn{\text{Cov}\left(Y_t^i,Y_t^j\right);}
\item \code{cov_jik_mfou} calculates the covariance between component i at time \eqn{t} and
component j at time \eqn{t+k}, \deqn{\text{Cov}\left(Y_t^i,Y_t^j\right);}
\eqn{k\in\mathbb{R}}.
}
}
