% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_est_gmm.R
\name{cov_mom_nw}
\alias{cov_mom_nw}
\title{Newey-West estimator of the asymptotic covariance matrix of moment conditions}
\usage{
cov_mom_nw(theta, my, type = "exa", bn = NULL, verbose = TRUE)
}
\arguments{
\item{theta}{Vector of coefficients.}

\item{my}{Numeric matrix containing the time series of the multivariate system as columns.}

\item{type}{One among \code{exa}, \code{asy}, \code{cau} to determine which version of the model the user wants to estimate
among exact, asymptotic, and causal.}

\item{bn}{Bandwidth parameter. If NULL, it is automatically selected.}

\item{verbose}{Logical. If TRUE, the automatically selected bandwidth is printed.}
}
\value{
Numeric matrix.
}
\description{
This function estimates the asymptotic covariance matrix of the moment conditions,
following the methodology outlined in Newey-West (1987, 1994). The moment conditions
are calculated for lags \eqn{\mathcal{L}=(0,1,2,3,5,20,50)} as is done in the paper.
}
\references{
Newey, Whitney K., and Kenneth D. West. “A Simple, Positive Semi-Definite, Heteroskedasticity and Autocorrelation Consistent Covariance Matrix.” Econometrica, vol. 55, no. 3, 1987, pp. 703–08. JSTOR, https://doi.org/10.2307/1913610. Accessed 22 July 2025.
Newey, Whitney K., and Kenneth D. West. "Automatic lag selection in covariance matrix estimation." The review of economic studies 61.4 (1994): 631-653.
}
