% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_est_gmm.R
\name{errors_gmm}
\alias{errors_gmm}
\title{Moment conditions}
\usage{
errors_gmm(
  theta,
  my,
  type = "exa",
  ts = FALSE,
  lag_max = 5,
  lag_add = c(20, 50),
  delta = 1/252
)
}
\arguments{
\item{theta}{Vector of the parameters.}

\item{my}{Numeric matrix containing the time series of the multivariate system as columns.}

\item{type}{One among \code{exa}, \code{asy}, \code{cau} to determine which version of the model the user wants to estimate
among exact, asymptotic, and causal.}

\item{ts}{Logical. If TRUE, a time series of length n - 49 is returned, where n
is the number of rows in my. If FALSE, average values are returned.}
}
\value{
Numeric matrix or vector, depending on the input to ts.
}
\description{
This function computes the moment conditions of the GMM estimator, i.e. the
differences between empirical and theoretical cross-covariances, \eqn{\hat\gamma-\gamma(\theta)}.
It uses default lags \eqn{\mathcal{L}=(0,1,2,4,5,20,50)}, in line with the paper specification.
}
