% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_a_fou}
\alias{est_a_fou}
\title{Estimator of \eqn{\alpha}}
\usage{
est_a_fou(y, H, nu)
}
\arguments{
\item{y}{Numeric vector representing a time series.}

\item{H}{Hurst coefficient or its estimate.}

\item{nu}{Diffusion coefficient or its estimate.}
}
\value{
Numeric value.
}
\description{
This function calculates the method of moment estimator of the speed of mean
reversion coefficient in a fractional Ornstein-Uhlenbeck process.
}
\references{
Wang, Xiaohu, Weilin Xiao, and Jun Yu. "Modeling and forecasting realized volatility with the fractional Ornstein????????Uhlenbeck process." Journal of Econometrics 232.2 (2023): 389-415.
}
