% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_rho_eta_mfou}
\alias{est_rho_eta_mfou}
\alias{est_asy_rho_eta_mfou}
\title{Method of moment estimator for the matrices \eqn{\rho} and \eqn{\eta} of the mfOU process}
\usage{
est_rho_eta_mfou(my, a, nu, H, k = 1, delta = 1/252)

est_asy_rho_eta_mfou(my, nu, H, k = 1, delta = 1/252)
}
\arguments{
\item{my}{Numeric matrix containing the time series in its columns.}

\item{a}{Numeric vector containing the speed of mean reversion coefficients.}

\item{nu}{Numeric vector containing the diffusion coefficients.}

\item{H}{Numeric vector containing the Hurst coefficients.}

\item{k}{lag for the covariance function that is inverted, \eqn{k>0}.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}
}
\value{
List of two numeric matrices.
}
\description{
These functions calculate the method of moment estimator of the parameter
matrices \eqn{\rho} and \eqn{\eta} characterizing the cross-covariance function
of the multivariate fractional Ornstein-Uhlenbeck process. These are
obtained by inverting the cross-covariances at lag \eqn{k} and \eqn{-k}.
\itemize{
\item \code{est_rho_eta_mfou} uses exact conditions;
\item \code{est_asy_rho_eta_mfou} uses conditions from the approximate cross-covariance
function in the regime of small speed of mean reversion.
}
}
\references{
Dugo, Ranieri, Giacomo Giorgio, and Paolo Pigato. "The multivariate fractional Ornstein-Uhlenbeck process." arXiv preprint arXiv:2408.03051 (2024).
}
