% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_gmm_mfou}
\alias{est_gmm_mfou}
\title{Generalized Method of Moments estimation of the mfOU process}
\usage{
est_gmm_mfou(
  my,
  lag_max = 5,
  lag_add = c(20, 50),
  delta = 1/252,
  type = "exa",
  step2 = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{my}{Numeric matrix containing the time series of the multivariate system in its columns.}

\item{type}{One among \code{exa}, \code{asy}, \code{cau} to determine which version of the model the user wants to estimate
among exact, asymptotic, and causal.}

\item{step2}{Logical. If TRUE, 2-step GMM is implemented. If FALSE, only the first step with the weighting
matrix given by the identity matrix weights is performed.}

\item{verbose}{Logical. If TRUE, optimization details are printed for each iteration.}
}
\value{
List of numeric objects containing estimated parameters.
}
\description{
This function estimates the parameters of a multivare fractional Ornstein-Uhlenbeck process
of dimension \eqn{d}, say \eqn{\theta}, by leveraging on the Generalized Method of Moments (GMM) methodology.
The GMM consists in minimizing a square distance between sample estimates of cross-covariances,
\eqn{\hat\gamma}, and theoretical ones, \eqn{\gamma(\theta)}. The vectors \eqn{\hat\gamma}
and \eqn{\gamma(\theta)} include cross-covariances evaluated for all combinations
\eqn{i,j=1,\dots,d} and all lags \eqn{k\in\mathcal{L}}.
The loss function being minimized is
\deqn{\left(\hat\gamma-\gamma(\theta)\right)^T W_n \left(\hat\gamma-\gamma(\theta)\right),}
for a positive semi-definite matrix \eqn{W_n}.
The methodology considers lags \eqn{\mathcal{L}=(0,1,2,3,5,20,50)} as in the paper.
}
\examples{
# Simulated data
mfou <-
  sim_mfou(n = 10000, delta = 1/252, m = 1,
           H = c(0.1, 0.2, 0.3),
           a = c(1, 0.5, 1),
           nu = c(1, 0.5, 1),
           mu = c(0, 0, 0),
           rho = matrix(c(1, 0.6, 0.7, 0.6, 1, 0.8, 0.7, 0.8, 1),
                        ncol = 3),
           eta = matrix(c(0, -0.1, -0.2, 0.1, 0, -0.3, 0.2, 0.3, 0),
                     ncol = 3))[2000 : 10000, ]
est_gmm_mfou(mfou)

# Log-realized volatilities
my = t(t(lrv[, c(2, 3)]) - apply(lrv[, c(2, 3)], 2, mean))
est_gmm_mfou(my)
}
\references{
Dugo, Ranieri, Giacomo Giorgio, and Paolo Pigato. "Multivariate Rough Volatility." arXiv preprint arXiv:2412.14353 (2024).
}
