% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_H_fou}
\alias{est_H_fou}
\title{Estimator of H}
\usage{
est_H_fou(y, verbose = FALSE)
}
\arguments{
\item{y}{Numeric vector representing a time series.}

\item{verbose}{Logical. If TRUE, warnings are printed if \eqn{H<0} or \eqn{H>1}.}
}
\value{
Numeric value.
}
\description{
This function calculates the change of frequency estimator (COF) of the Hurst
coefficient of a fractional Ornstein-Uhlenbeck process. The result is restricted between
\eqn{10^{-3}} and \eqn{1 - 10^{-3}} in case of finite sample degeneracies.
}
\references{
Wang, Xiaohu, Weilin Xiao, and Jun Yu. "Modeling and forecasting realized volatility with the fractional Ornstein????????Uhlenbeck process." Journal of Econometrics 232.2 (2023): 389-415.
}
