% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_rhoij_etaij_mfou}
\alias{est_rhoij_etaij_mfou}
\alias{est_asy_rhoij_etaij_mfou}
\title{Method of moment estimator for \eqn{\rho} and \eqn{\eta} of the mfOU process}
\usage{
est_rhoij_etaij_mfou(yi, yj, ai, aj, nui, nuj, Hi, Hj, k = 1, delta = 1/252)

est_asy_rhoij_etaij_mfou(yi, yj, nui, nuj, Hi, Hj, k = 1, delta = 1/252)
}
\arguments{
\item{yi}{Numeric vector representing a time series.}

\item{yj}{Numeric vector of the same length of yi representing a time series.}

\item{ai}{Speed of mean reversion of component i.}

\item{aj}{Speed of mean reversion of component j.}

\item{nui}{Diffusion coefficient of component i.}

\item{nuj}{Diffusion coefficient of component j.}

\item{Hi}{Hurst coefficient of component i.}

\item{Hj}{Hurst coefficient of component j.}

\item{k}{lag for the covariance function that is inverted, \eqn{k>0}.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}
}
\value{
Numeric vector.
}
\description{
These functions calculates the method of moment estimator of the parameters
\eqn{\rho_{i,j}} and \eqn{\eta_{i,j}} characterizing the cross-covariance
function of the multivariate fractional Ornstein-Uhlenbeck process. It is
obtained by inverting the cross-covariances at lags \eqn{k} and \eqn{-k}.
\itemize{
\item \code{est_rhoij_etaij_mfou} uses exact conditions;
\item \code{est_asy_rhoij_etaij_mfou} uses conditions from the approximate cross-covariance
function in the regime of small speed of mean reversion.
}
}
\references{
Dugo, Ranieri, Giacomo Giorgio, and Paolo Pigato. "The multivariate fractional Ornstein-Uhlenbeck process." arXiv preprint arXiv:2408.03051 (2024).
}
