% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfbm.R
\name{est_H_fbm}
\alias{est_H_fbm}
\alias{est_sig_fbm}
\alias{est_rhoij_mfbm}
\alias{est_etaij_mfbm}
\alias{est_rho_eta_mfbm}
\title{Method of moment estimation of the mfBm}
\usage{
est_H_fbm(x)

est_sig_fbm(x, H = NULL, delta = 1)

est_rhoij_mfbm(xi, xj)

est_etaij_mfbm(xi, xj)

est_rho_eta_mfbm(mx)
}
\arguments{
\item{x}{Numeric vector representing a time series.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years).}

\item{xi}{Numeric vector representing a time series.}

\item{xj}{Numeric vector of the same length of xj representing a time series.}

\item{mx}{Numeric matrix containing the time series in its columns.}
}
\value{
Numeric value, except for \code{est_rho_eta_mfbm} that returns a list of
numeric matrices.
}
\description{
Method of moment estimator of the multivariate fractional Brownian motion.
\itemize{
\item \code{est_H_fbm} estimates the Hurst coefficient of a fractional Brownian motion;
\item \code{est_sig_fbm}  estimates the scale coefficient of a fractional Brownian
motion (also its variance at \eqn{t=1});
\item \code{est_rhoij_mfbm} estimates the (contemporaneous) correlation coefficient between
two fractional Brownian motions;
\item \code{est_etaij_mfbm} estimates the asymmetry coefficient between
two fractional Brownian motions;
\item \code{est_rho_eta_mfbm} combines the two functions above to estimates the matrices
\eqn{\rho} and \eqn{\eta} for a multivariate fractional Brownian motion of
arbitrary dimension.
}
}
\references{
Bibinger, Markus, Jun Yu, and Chen Zhang. "Modeling and Forecasting Realized Volatility with Multivariate Fractional Brownian Motion." arXiv preprint arXiv:2504.15985 (2025).
}
