% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_mm_fou}
\alias{est_mm_fou}
\title{Two-step estimator of fOU}
\usage{
est_mm_fou(y, delta = 1/252, H = NULL)
}
\arguments{
\item{y}{Numeric vector representing a time series.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}

\item{H}{Optional parameter corresponding to the Hurst coefficient. If provided,
the procedure is one-step.}
}
\value{
Numeric vector.
}
\description{
This function calculates the two-step estimator of the parameters of a
fractional Ornstein-Uhlenbeck (fOU) process. It combines the Change of
Frequency estimator of H with a method of moment estimator for the remaining
parameters \eqn{\nu}, \eqn{\alpha}, and \eqn{\mu}.
}
\references{
Wang, Xiaohu, Weilin Xiao, and Jun Yu. "Modeling and forecasting realized volatility with the fractional Ornstein????????Uhlenbeck process." Journal of Econometrics 232.2 (2023): 389-415.
}
