% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_mm_mfou}
\alias{est_mm_mfou}
\title{Two-step estimator of the mfOU process}
\usage{
est_mm_mfou(my, delta = 1/252, k = 1, type = "exa")
}
\arguments{
\item{my}{Numeric matrix containing the time series of the multivariate system as columns.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}

\item{k}{lag for the covariance function that is inverted, \eqn{k>0}.}

\item{type}{One between \code{"exa"} and \code{"asy"} to determine whether the exact cross-covariance condition
or the asymptotic one is to be used.}
}
\value{
List of numeric matrices
}
\description{
This function calculates the method of moment estimator of all the parameters
of a multivariate fractional Ornstein-Uhlenbeck process of dimension \eqn{d}.
The parameters govering the univariate marginal distributions,
\eqn{H_i,\ \nu_i,\ \alpha_i,\ \text{and}\ \mu_i,\ i=1,\dots,d}, are obtained in a
first step, whereas those determining the joint dynamics, \eqn{\rho\ \text{and}\ \eta},
are obtained pairwise using the estimates from the previous step as input.
}
\references{
Dugo, Ranieri, Giacomo Giorgio, and Paolo Pigato. "The multivariate fractional Ornstein-Uhlenbeck process." arXiv preprint arXiv:2408.03051 (2024).

Wang, Xiaohu, Weilin Xiao, and Jun Yu. "Modeling and forecasting realized volatility with the fractional Ornstein????????Uhlenbeck process." Journal of Econometrics 232.2 (2023): 389-415.
}
