% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_mu_fou}
\alias{est_mu_fou}
\title{Estimator of \eqn{\mu}}
\usage{
est_mu_fou(y)
}
\arguments{
\item{y}{Numeric vector representing a time series.}
}
\value{
Numeric value.
}
\description{
This function calculates the method of moment estimator of the long-term mean
of a fractional Ornstein-Uhlenbeck process. It corresponds to a sample average.
}
\references{
Wang, Xiaohu, Weilin Xiao, and Jun Yu. "Modeling and forecasting realized volatility with the fractional Ornstein????????Uhlenbeck process." Journal of Econometrics 232.2 (2023): 389-415.
}
