% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mfou.R
\name{est_nu_fou}
\alias{est_nu_fou}
\title{Estimator of \eqn{\nu}}
\usage{
est_nu_fou(y, H, delta = 1/252)
}
\arguments{
\item{y}{Numeric vector representing a time series.}

\item{H}{Hurst coefficient or its estimate.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}
}
\value{
Numeric value.
}
\description{
This function calculates the method of moment estimator of the diffusion
coefficient of a fractional Ornstein-Uhlenbeck process.
}
\references{
Wang, Xiaohu, Weilin Xiao, and Jun Yu. "Modeling and forecasting realized volatility with the fractional Ornstein????????Uhlenbeck process." Journal of Econometrics 232.2 (2023): 389-415.
}
