% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfbm.R
\name{etaij_caus}
\alias{etaij_caus}
\alias{eta_caus}
\title{Expression of \eqn{\eta_{i,j}} in the causal mfBm}
\usage{
etaij_caus(rhoij, Hi, Hj)

eta_caus(rho, H)
}
\arguments{
\item{rhoij}{Correlation coefficient between components i and j.}

\item{Hi}{Hurst coefficient of component i.}

\item{Hj}{Hurst coefficient of component j.}

\item{rho}{Symmetric matrix of instantaneous correlation coefficients.}

\item{H}{Vector of Hurst coefficients.}
}
\value{
Numeric value for \code{etaij_caus} or matrix for \code{eta_caus}.
}
\description{
Formulas for the computation of the asymmetry matrix \eqn{\eta},
respectively  coefficient \eqn{\eta_{i,j}},
as a function of the other parameters, \eqn{\rho} and \eqn{H}, respectively
\eqn{\rho_{i,j},\ H_i, \text{and}\ H_j}, in the casusal multivariate fractional
Brownian motion. The causal case corresponds to the process depending only on
past innovations in the moving average representation.
}
\references{
Amblard, Pierre-Olivier, et al. "Basic properties of the multivariate fractional Brownian motion." arXiv preprint arXiv:1007.0828 (2010).
}
