% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{cov_jhdt_mfbm}
\alias{cov_jhdt_mfbm}
\alias{cov_dt_mfbm}
\alias{cov_0_mfbm}
\alias{fcast_point_mfbm}
\alias{fcast_se_mfbm}
\title{Forecasting the mfBm using Gaussian conditioning}
\usage{
cov_jhdt_mfbm(n, h, H, sig, rho, eta, j, delta)

cov_dt_mfbm(n, H, sig, rho, eta, delta)

cov_0_mfbm(n, H, sig, rho, eta, delta)

fcast_point_mfbm(my, h, H, sig, rho, eta, delta, mu = NULL, plot = FALSE)

fcast_se_mfbm(n, h, H, sig, rho, eta, delta)
}
\arguments{
\item{n}{Integer. Length of the time series.}

\item{h}{Integer. Forecast horizon (e.g. 1, 5, 10 steps ahead).}

\item{H}{Vector of Hurst coefficients.}

\item{sig}{Vector of scale coefficients.}

\item{rho}{Symmetric matrix of instantaneous correlation coefficients.}

\item{eta}{Antisymmetric matrix of asymmetry coefficients.}

\item{j}{Component of the system. Integer between 1 and \eqn{d} or column name.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years).}

\item{my}{Numeric matrix containing the time series in its columns.}

\item{mu}{Vector of means. Default is NULL, in which case it is computed as sample average.}

\item{plot}{Logical. Default is FALSE. If TRUE, a panel of plots will exhibit
the weights of the linear combination.}
}
\value{
Numeric matrices.
}
\description{
Using a standard result for Gaussian random vectors, these functions allow
us to compute point forecasts and associated standard errors in the framework
of the multivariate fractional Brownian motion of dimension \eqn{d}.

Given that the vector that we want to predict, \eqn{X:=B_{(t+h)\Delta}\in\mathbb{R}^d}, and the
history of information, \eqn{Y:=\left(B_{t\Delta},\dots, B_{\Delta}\right)\in\mathbb{R}^{d\times t}},
are jointly Gaussian, we use that:

\deqn{\mu_{X|Y}=\mu_X+\Sigma_{XY}\Sigma_{YY}^{-1}\left(Y-\mu_Y\right),}
\deqn{\Sigma_{XX|Y}=\Sigma_{XX}-\Sigma_{XY}\Sigma_{YY}^{-1}\Sigma_{XY}^T,}
where we used the notation \eqn{\mu_Z=\mathbb{E}(Z)} and
\eqn{\Sigma_{ZU} = \mathbb{E}((Z-\mu_Z)(U-\mu_U)^T)}.
\itemize{
\item \code{cov_jhdt_mfbm} calculates the matrix \eqn{\Sigma_{XY}} described above;
\item \code{cov_dt_mfbm} calculates the matrix \eqn{\Sigma_{YY}} described above;
\item \code{cov_0_mfbm} calculates the matrix \eqn{\Sigma_{XX}} described above;
\item \code{fcast_point_mfbm} calculates point forecasts, i.e. the vector \eqn{\mu_{X|Y}} described above;
\item \code{fcast_se_mfbm} calculates standard errors of the forecasts, i.e. the matrix \eqn{\Sigma_{XX|Y}} described above;
}

The functions \code{fcast_point_mfbm} and \code{fcast_se_mfbm} are not efficient in
rolling window applications with fixed parameters. In which case,
efficiency gains could often be obtained by avoiding redundant matrix operations.
}
\references{
Bibinger, Markus, Jun Yu, and Chen Zhang. "Modeling and Forecasting Realized Volatility with Multivariate Fractional Brownian Motion." arXiv preprint arXiv:2504.15985 (2025).

https://en.wikipedia.org/wiki/Multivariate_normal_distribution
}
