% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_mfou.R
\name{gof_fou}
\alias{gof_fou}
\alias{gof_asy_fou}
\title{Goodness of covariance fit for fOU}
\usage{
gof_fou(y, est, lag_max = 50, delta = 1/252)

gof_asy_fou(y, est, lag_max = 50, delta = 1/252)
}
\arguments{
\item{y}{Numeric vector representing the time series.}

\item{est}{Vector of estimated parameters. In the order: Hurst coefficient,
diffusion coefficient, and speed of mean reversion (the latter is unnecessary
for the asymptotic version) or simply output from \code{est_mm_fou}}

\item{lag_max}{Positive integer. All the lags between 0 (included) and \code{lag_max} are considered.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}
}
\value{
Plot and numeric matrix containing the theoretical and empirical cross-covariances.
}
\description{
These functions evaluate (graphically) the goodness of fit of estimated
parameters of a fractional Ornstein-Uhlenbeck process in terms of the distance
between theoretical (model) and empirical (data) autocovariances.
\itemize{
\item \code{gof_fou} considers exact autocovariances;
\item \code{gof_asy_fou} considers asymptotic autocovariances arising in the regime
of small speed of mean reversion.
}
}
\examples{
y = lrv[, 2] - mean(lrv[, 2])
est = est_mm_fou(y)
gof_fou(y, est) # exact version
gof_asy_fou(y, est) # asymptotic version
}
