% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_mfou.R
\name{gof_mfou}
\alias{gof_mfou}
\alias{gof_asy_mfou}
\title{Goodness of covariance fit for mfOU}
\usage{
gof_mfou(my, est, si = 1, sj = 2, lag_max = 50, delta = 1/252)

gof_asy_mfou(my, est, si = 1, sj = 2, lag_max = 50, delta = 1/252)
}
\arguments{
\item{my}{Numeric matrix containing the time series of the multivariate system as columns.}

\item{est}{List of estimated parameters, possibly from \code{est_gmm_mfou} called on `my.}

\item{si}{Index identifying the component i of the system (or its column name in my)}

\item{sj}{Index identifying the component j of the system (or its column name in my)}

\item{lag_max}{Positive integer. All the lags between 0 (included) and \code{lag_max} are considered.}

\item{delta}{Time step between consecutive observations in the time series y (e.g., 1/252 for daily data if time is in years), \eqn{\delta\in\mathbb{R}}.}
}
\value{
Numeric matrix containing the theoretical and empirical cross-covariances and plots.
}
\description{
These functions evaluate (graphically) the goodness of fit of estimated
parameters of a multivariate fractional Ornstein-Uhlenbeck process in
terms of the distance between theoretical (model) and empirical (data)
cross-covariances in a pairwise fashion.
\itemize{
\item \code{gof_mfou} considers exact cross-covariances;
\item \code{gof_asy_mfou} considers asymptotic cross-covariances arising in the regime
of small speed of mean reversion.
}
}
\examples{
# exact version
my = t(t(lrv[, c(2, 3)]) - apply(lrv[, c(2, 3)], 2, mean))
est = est_gmm_mfou(my)
gof_mfou(my, est)

# asymptotic version
my = t(t(lrv[, c(2, 3)]) - apply(lrv[, c(2, 3)], 2, mean))
est = est_gmm_mfou(my, type = "asy")
gof_asy_mfou(my, est)
}
