% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfou.R
\name{I_ij}
\alias{I_ij}
\alias{I_ji}
\title{Integral appearing in the cross-covariance function}
\usage{
I_ij(ai, aj, Hi, Hj, k)

I_ji(ai, aj, Hi, Hj, k)
}
\arguments{
\item{ai}{Speed of mean reversion of component i, \eqn{\alpha_i}.}

\item{aj}{Speed of mean reversion of component j, \eqn{\alpha_j}.}

\item{Hi}{Hurst coefficient of component i, \eqn{H_i}.}

\item{Hj}{Hurst coefficient of component j, \eqn{H_j}.}

\item{k}{Lag, \eqn{k > 0}.}
}
\value{
Numeric
}
\description{
Integral appearing in the cross-covariance function
}
