% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_est_gmm.R
\name{jacobian_gmm}
\alias{jacobian_gmm}
\title{Jacobian of the moment conditions}
\usage{
jacobian_gmm(
  theta,
  my,
  type = "exa",
  lag_max = 5,
  lag_add = c(20, 50),
  delta = 1/252
)
}
\arguments{
\item{theta}{Vector of the parameters.}

\item{my}{Numeric matrix containing the time series of the multivariate system as columns.}

\item{type}{One among \code{exa}, \code{asy}, \code{cau} to determine which version of the model the user wants to estimate
among exact, asymptotic, and causal.}
}
\value{
Numeric matrix.
}
\description{
This function calculates the Jacobian of the moment conditions, given by the difference
between empirical and theoretical cross-covariances, \eqn{\hat\gamma-\gamma(\theta)}.
}
