% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_mfou.R
\name{lin_acf_fou}
\alias{lin_acf_fou}
\alias{lin_ccf_mfou}
\title{Check for slow mean reversion}
\usage{
lin_acf_fou(y, H = NULL, lag_max = 50)

lin_ccf_mfou(yi, yj, Hi = NULL, Hj = NULL, lag_max = 50)
}
\arguments{
\item{y}{Numeric vector representing a time series.}

\item{H}{Hurst coefficient. Default is NULL, in which case H is obtained with the COF estimator.}

\item{lag_max}{Positive integer. All the lags between 0 (included) and \code{lag_max} are considered.}

\item{yi}{Numeric vector representing a time series.}

\item{yj}{Numeric vector of the same length of yi representing a time series.}

\item{Hi}{Hurst coefficient of component i. Default is NULL, in which case Hi is obtained with the COF estimator.}

\item{Hj}{Hurst coefficient of component j. Default is NULL, in which case Hj is obtained with the COF estimator.}
}
\value{
Numeric matrices containing theoretical and empirical covariances.
}
\description{
These functions verify (graphically) the regime of slow mean reversion, which consists
in small speed of mean reversion coefficients with respect to the time scale considered.
This regime is characterized by the linearity of the autocovariances and
cross-covariances when considered against a suitable power of the lags,
determined by the Hurst coefficients.
\itemize{
\item \code{lin_acf_fou} evaluates the condition on the autocovariance of a univariate component;
\item \code{lin_ccf_mfou} evaluates the condition on the cross-covariances between two components;
}

The conditions should be considered for all components and pairs of components
before modeling the entire system in the asymptotic regime.
}
\examples{
# univariate case
y = lrv[, 2] - mean(y = lrv[, 2])
lin_acf_fou(y)

# multivariate case
yi = lrv[, 1]
yj = lrv[, 2]
lin_ccf_mfou(yi, yj)
}
