% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{lrv}
\alias{lrv}
\title{Log-realized volatility}
\format{
A data frame with 7309 rows and 8 columns:
\describe{
\item{date}{Date of observation.}
\item{AAPL}{log-realized vol for Apple.}
\item{AMGN}{log-realized vol for Amgen.}
\item{ALD}{log-realized vol for Allied Signal.}
\item{AXP}{log-realized vol for American Express.}
\item{BA}{log-realized vol for Boeing.}
\item{CAT}{log-realized vol for Caterpillar.}
}
}
\source{
Data obtained from Risk Lab (https://dachxiu.chicagobooth.edu/#risklab).
}
\usage{
lrv
}
\description{
A dataset containing the logarithm of a number of realized volatilities
(annualized and tranformed to percentages). This dataset is used in Bibinger,
Yu, and Zhang (2025). It is not the same as in the paper Dugo, Giorgio, and
Pigato (2025) because of not having permission to disseminate it.
}
\keyword{datasets}
