% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers.R
\name{nfevd}
\alias{nfevd}
\title{Normalized forecast error variance decomposition}
\usage{
nfevd(H, rho)
}
\arguments{
\item{H}{Vector of Hurst coefficients.}

\item{rho}{Matrix of instantaneous correlation coefficients.}
}
\value{
A numeric matrix
}
\description{
Normalized forecast error variance decomposition
}
\references{
Dugo, Ranieri, Giacomo Giorgio, and Paolo Pigato. "Multivariate Rough Volatility." arXiv preprint arXiv:2412.14353 (2024).

Pesaran, H. Hashem, and Yongcheol Shin. "Generalized impulse response analysis in linear multivariate models." Economics letters 58.1 (1998): 17-29.
}
