% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers.R
\name{spillovers}
\alias{spillovers}
\title{Spillover indices}
\usage{
spillovers(H, rho, plot = FALSE)
}
\arguments{
\item{H}{Vector of Hurst coefficients.}

\item{rho}{Matrix of diffusion correlation coefficients.}

\item{plot}{Logical. If TRUE, two plots showing directional and net
pairwise spillovers are printed}
}
\value{
List of numeric objects.
}
\description{
This function computes spillover indices in the
multivariate fractional Ornstein-Uhlenbeck process and multivariate fractional
Brownian motion models of size \eqn{d} starting
from (causal) estimates for the parameters \eqn{H} and \eqn{\rho}.
}
\examples{
my = rv[, c("FTSE", "N225", "SPX", "SSEC")]

# mfOU
est = est_mde_mfou(my, type = "cau")
spillovers(est$univ[1, ], est$rho, plot = TRUE)

# mfBm
H = apply(my, 2, est_H_fbm)
rho = est_rho_eta_mfbm(my)$rho
spillovers(H, rho, plot = TRUE)
}
\references{
Dugo, Ranieri, Giacomo Giorgio, and Paolo Pigato. "Multivariate Rough Volatility." arXiv preprint arXiv:2412.14353 (2024).

Diebold, Francis X., and Kamil Yilmaz. "Better to give than to receive: Predictive directional measurement of volatility spillovers." International Journal of forecasting 28.1 (2012): 57-66.
}
