% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfbm.R
\name{var_fbm}
\alias{var_fbm}
\title{Variance function of the fBm}
\usage{
var_fbm(t, H, sigma = 1)
}
\arguments{
\item{t}{Time.}

\item{H}{Hurst coefficient.}

\item{sigma}{Scale coefficient (square root of variance at \eqn{t=1}).}
}
\value{
Numeric value.
}
\description{
Variance function of the fractional Brownian motion, based on model parameters:
\deqn{\text{Var}(B_t^H).}
}
