% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mfou.R
\name{var_fou}
\alias{var_fou}
\title{Variance function of the fOU process}
\usage{
var_fou(a, nu, H)
}
\arguments{
\item{a}{Speed of mean reversion coefficient.}

\item{nu}{Diffusion coefficient.}

\item{H}{Hurst coefficient.}
}
\value{
Numeric value.
}
\description{
Variance of the fractional Ornstein-Uhlenbeck (fOU) process, based on model parameters,
\eqn{\text{Var}\left(Y_{t}\right)}.
}
